+++ src/utils/helpers/focus-trap/focus-trap.js
@@ -8,12 +8,23 @@ function blockTabbing(ev) {
 function trap(firstFocusableElement, lastFocusableElement, ev) {
   if (ev.key === 'Tab' || ev.keyCode === 9) {
     if (ev.shiftKey) /* shift + tab */ {
-      if (document.activeElement === firstFocusableElement) {
+      if (typeof firstFocusableElement === 'function') {
+        ev.preventDefault();
+        if (document.activeElement !== lastFocusableElement) {
+          lastFocusableElement.focus();
+        } else {
+          firstFocusableElement();
+        }
+      } else if (document.activeElement === firstFocusableElement) {
         lastFocusableElement.focus();
         ev.preventDefault();
       }
     } else if (document.activeElement === lastFocusableElement) {
-      firstFocusableElement.focus();
+      if (typeof firstFocusableElement === 'function') {
+        firstFocusableElement();
+      } else {
+        firstFocusableElement.focus();
+      }
       ev.preventDefault();
     }
   }
@@ -21,10 +32,14 @@ function trap(firstFocusableElement, lastFocusableElement, ev) {
 // eslint-disable-next-line max-len
 const defaultFocusableSelectors = 'button, [href], input:not([type="hidden"]), select, textarea, [tabindex]:not([tabindex="-1"])';
 
-const focusTrap = (element, fs = defaultFocusableSelectors) => {
-  const focusableElements = element.querySelectorAll(fs);
-  const firstFocusableElement = focusableElements[0];
+const focusTrap = (element, focusFirstElement) => {
+  const focusableElements = element.querySelectorAll(defaultFocusableSelectors);
+  let firstFocusableElement = focusableElements[0];
   const lastFocusableElement = focusableElements[focusableElements.length - 1];
+
+  if (focusFirstElement) {
+    firstFocusableElement = focusFirstElement;
+  }
 
   if (focusableElements.length <= 0) {
     document.addEventListener('keydown', blockTabbing);